Created a separate component called FilterSelects
